movies_admin/requirements/base.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.5.6
black==21.4b2
click==7.1.2
colorama==0.4.4
Django==3.1
django-extensions==3.1.3
django-model-utils==4.0.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
pylint==2.8.2
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
wrapt==1.12.1
factory-boy==3.2.0
Faker==8.1.1
whitenoise==5.2.0